Fix some wrong integer types

+++ common.c
@@ -91,7 +91,7 @@ RRStatus checkLeader(RedisRaftCtx *rr, RaftReq *req, Node **ret_leader)
                 return RR_OK;
             }
 
-            int reply_maxlen = strlen(leader_node->addr.host) + 15;
+            size_t reply_maxlen = strlen(leader_node->addr.host) + 15;
             char *reply = RedisModule_Alloc(reply_maxlen);
             snprintf(reply, reply_maxlen, "MOVED %s:%u", leader_node->addr.host, leader_node->addr.port);
             RedisModule_ReplyWithError(req->ctx, reply);
@@ -175,7 +175,7 @@ static void raftize_commands(RedisModuleCommandFilterCtx *filter)
     /* Don't process any excluded command */
     char **c = excluded_commands;
     while (*c != NULL) {
-        int len = strlen(*c);
+        size_t len = strlen(*c);
         if (cmdname_len == len && !strncasecmp(cmdname, *c, len)) {
             return;
         }
